test: migrate node specs to main
