+++ spec-main/webview-spec.ts
@@ -3,8 +3,8 @@ import * as url from 'url';
 import { BrowserWindow, session, ipcMain, app, WebContents } from 'electron/main';
 import { closeAllWindows } from './window-helpers';
 import { emittedOnce, emittedUntil } from './events-helpers';
-import { ifit, delay, defer } from './spec-helpers';
-import { AssertionError, expect } from 'chai';
+import { ifit, delay, defer, itremote, useRemoteContext } from './spec-helpers';
+import { expect } from 'chai';
 import * as http from 'http';
 import { AddressInfo } from 'net';
 
@@ -49,24 +49,6 @@ async function loadWebViewAndWaitForMessage (w: WebContents, attributes: Record<
   const { message } = await loadWebViewAndWaitForEvent(w, attributes, 'console-message');
   return message;
 };
-
-async function itremote (name: string, fn: Function, args?: any[]) {
-  it(name, async () => {
-    const w = new BrowserWindow({ show: false, webPreferences: { nodeIntegration: true, contextIsolation: false, webviewTag: true } });
-    defer(() => w.close());
-    w.loadURL('about:blank');
-    const { ok, message } = await w.webContents.executeJavaScript(`(async () => {
-      try {
-        const chai_1 = require('chai')
-        await (${fn})(...${JSON.stringify(args ?? [])})
-        return {ok: true};
-      } catch (e) {
-        return {ok: false, message: e.message}
-      }
-    })()`);
-    if (!ok) { throw new AssertionError(message); }
-  });
-}
 
 describe('<webview> tag', function () {
   const fixtures = path.join(__dirname, '..', 'spec', 'fixtures');
@@ -1105,6 +1087,7 @@ describe('<webview> tag', function () {
     });
 
     describe('preload attribute', () => {
+      useRemoteContext({ webPreferences: { webviewTag: true } });
       it('loads the script before other scripts in window', async () => {
         const message = await loadWebViewAndWaitForMessage(w, {
           preload: `${fixtures}/module/preload.js`,
@@ -1431,6 +1414,7 @@ describe('<webview> tag', function () {
   });
 
   describe('events', () => {
+    useRemoteContext({ webPreferences: { webviewTag: true } });
     let w: WebContents;
     before(async () => {
       const window = new BrowserWindow({
